Fix product api

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/interactionapi/request/AddProductToWarehouseRequest.java
@@ -14,7 +14,6 @@ import java.util.UUID;
 public class AddProductToWarehouseRequest {
     @NotNull
     UUID productId;
-    
     @NotNull
-    int quantity;
+    Long quantity;
 }
